requirements.txt
# Flask Web Framework
Flask==2.3.3

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
PyMySQL==1.1.0

# Authentication & Security
Flask-Login==0.6.3
Werkzeug==2.3.7

# Forms
Flask-WTF==1.1.1
WTForms==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2

# CSV Processing
# (標準ライブラリのcsvモジュールを使用)

# File Upload & Processing
# (標準ライブラリのtempfile, os, ioモジュールを使用)

# HTTP & Web
# (標準ライブラリのdatetime, localeモジュールを使用)

# Development & Debug
# (標準ライブラリのlogging, sysモジュールを使用) 